massive refactor
