+++ forum/post.py
@@ -1,0 +1,106 @@
+import datetime
+from flask import *
+from flask_login import login_required, current_user
+
+from forum.comment import Comment
+from forum.subforum import Subforum, generateLinkPath
+from forum.forum import db, app, valid_title, valid_content
+from forum.app import error
+
+
+class Post(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.Text)
+    content = db.Column(db.Text)
+    comments = db.relationship("Comment", backref="post")
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
+    subforum_id = db.Column(db.Integer, db.ForeignKey('subforum.id'))
+    postdate = db.Column(db.DateTime)
+
+    # cache stuff
+    lastcheck = None
+    savedresponce = None
+
+    def __init__(self, title, content, postdate):
+        self.title = title
+        self.content = content
+        self.postdate = postdate
+
+    def get_time_string(self):
+        # this only needs to be calculated every so often, not for every request
+        # this can be a rudamentary chache
+        now = datetime.datetime.now()
+        if self.lastcheck is None or (now - self.lastcheck).total_seconds() > 30:
+            self.lastcheck = now
+        else:
+            return self.savedresponce
+
+        diff = now - self.postdate
+
+        seconds = diff.total_seconds()
+        print(seconds)
+        if seconds / (60 * 60 * 24 * 30) > 1:
+            self.savedresponce = " " + str(int(seconds / (60 * 60 * 24 * 30))) + " months ago"
+        elif seconds / (60 * 60 * 24) > 1:
+            self.savedresponce = " " + str(int(seconds / (60 * 60 * 24))) + " days ago"
+        elif seconds / (60 * 60) > 1:
+            self.savedresponce = " " + str(int(seconds / (60 * 60))) + " hours ago"
+        elif seconds / (60) > 1:
+            self.savedresponce = " " + str(int(seconds / 60)) + " minutes ago"
+        else:
+            self.savedresponce = "Just a moment ago!"
+
+        return self.savedresponce
+
+
+@login_required
+@app.route('/addpost')
+def addpost():
+    subforum_id = int(request.args.get("sub"))
+    subforum = Subforum.query.filter(Subforum.id == subforum_id).first()
+    if not subforum:
+        return error("That subforum does not exist!")
+
+    return render_template("createpost.html", subforum=subforum)
+
+
+@app.route('/viewpost')
+def viewpost():
+    postid = int(request.args.get("post"))
+    post = Post.query.filter(Post.id == postid).first()
+    if not post:
+        return error("That post does not exist!")
+    if not post.subforum.path:
+        subforum.path = generateLinkPath(post.subforum.id)
+    comments = Comment.query.filter(Comment.post_id == postid).order_by(
+        Comment.id.desc())  # no need for scalability now
+    return render_template("viewpost.html", post=post, path=subforum.path, comments=comments)
+
+
+@login_required
+@app.route('/action_post', methods=['POST'])
+def action_post():
+    subforum_id = int(request.args.get("sub"))
+    subforum = Subforum.query.filter(Subforum.id == subforum_id).first()
+    if not subforum:
+        return redirect(url_for("subforums"))
+
+    user = current_user
+    title = request.form['title']
+    content = request.form['content']
+    # check for valid posting
+    errors = []
+    retry = False
+    if not valid_title(title):
+        errors.append("Title must be between 4 and 140 characters long!")
+        retry = True
+    if not valid_content(content):
+        errors.append("Post must be between 10 and 5000 characters long!")
+        retry = True
+    if retry:
+        return render_template("createpost.html", subforum=subforum, errors=errors)
+    post = Post(title, content, datetime.datetime.now())
+    subforum.posts.append(post)
+    user.posts.append(post)
+    db.session.commit()
+    return redirect("/viewpost?post=" + str(post.id))
